feat: S3 유틸리티 함수 추가 및 resizeImage 미들웨어에 적용
- S3 관련 작업을 담당하는 getFileFromS3, uploadResizedImage, deleteOriginalFile 함수를 utils/s3.ts에 새로 추가
- resizeImage 미들웨어에서 S3 관련 로직을 유틸리티 함수로 대체
